backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.7.0

